+++ test/public-types/smoke.tsx
@@ -1,9 +1,6 @@
 import 'react-color/es/styles/index.css';
 import 'react-color/es/styles/pickers/sketch.css';
 import 'react-color/es/styles/common/editable-input.css';
-import 'react-color/lib/styles/index.css';
-import 'react-color/lib/styles/pickers/chrome.css';
-import 'react-color/lib/styles/common/editable-input.css';
 
 import type { ComponentProps } from 'react';
 import React from 'react';
@@ -21,6 +18,7 @@ import ChromePickerDefault, {
   type RGBAColor,
 } from 'react-color';
 import SketchPickerEsm from 'react-color/es/Sketch';
+import HuePickerEsm from 'react-color/es/Hue';
 
 const rgba: RGBAColor = { r: 255, g: 0, b: 0, a: 1 };
 const hsl: HSLAColor = { h: 0, s: 1, l: 0.5, a: 1 };
@@ -69,6 +67,7 @@ const elements = (
     <SketchPicker {...sketchProps} theme="light" classNames={pickerClassNames} />
     <GooglePicker {...googleProps} />
     <SketchPickerEsm color="#00ff00" />
+    <HuePickerEsm color="#00ff00" />
     <EditableInput value="ff0000" onChange={() => undefined} />
     <WrappedPicker color="#ffffff" />
   </div>